Bring in ITextTypeProvider
This is a logical dependency for determining when to allow additional @
tokens for parsing CSS in Razor files.  The VS implementation of this
interface depends on the VS editor platform, which is a dependency we
are avoiding in this repo.

This commit addresses the compilation break, but we will need to provide
an implementation of this interface at some point.